backend/pyproject.toml
[project]
name = "hack-hatchery-backend"
version = "0.1.0"
description = "Website backend for hackhatchery.com"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "jinja2>=3.1.6",
    "markdown>=3.8.2",
    "markdown-extensions>=0.0.1",
    "pymdown-extensions>=10.16.1",
    "python-frontmatter>=1.1.0",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.35.0",
    "watchdog>=6.0.0",
]
